Ref: moving built-in logger to a dedicated file.
