+++ src/config-type.ts
@@ -2,8 +2,9 @@ import type compression from "compression";
 import { IRouter, Request, RequestHandler } from "express";
 import type fileUpload from "express-fileupload";
 import { ServerOptions } from "node:https";
+import { BuiltinLoggerConfig } from "./builtin-logger";
 import { AbstractEndpoint } from "./endpoint";
-import { AbstractLogger, ActualLogger, BuiltinLoggerConfig } from "./logger";
+import { AbstractLogger, ActualLogger } from "./logger";
 import { Method } from "./method";
 import { AnyResultHandlerDefinition } from "./result-handler";
 import { ListenOptions } from "node:net";
