+++ src/server.ts
@@ -2,8 +2,9 @@ import express from "express";
 import type compression from "compression";
 import http from "node:http";
 import https from "node:https";
+import { BuiltinLogger } from "./builtin-logger";
 import { AppConfig, CommonConfig, ServerConfig } from "./config-type";
-import { BuiltinLogger, isLoggerInstance } from "./logger";
+import { isLoggerInstance } from "./logger";
 import { loadPeer } from "./peer-helpers";
 import { defaultResultHandler } from "./result-handler";
 import { Parsers, Routing, initRouting } from "./routing";
